examples/openai/requirements.txt
fastapi[all]
gguf
jinja2
jsonargparse
pydantic
sse-starlette
uvicorn[all]